scripts/requirements.txt
html-sanitizer
pillow
requests
tqdm
yarl